tidy the route package directory
